Make AllergyId and AllergyName properties nullable
Updated the `AllergyId` property in the `AllergyUpdateDto` class to be a nullable `int?` instead of a non-nullable `int`. Similarly, updated the `AllergyName` property to be a nullable `string?` instead of a non-nullable `string`.

These changes provide greater flexibility by allowing `null` values for these properties, which may be useful in scenarios where these fields are optional.